+++ src/Universum/Foldable.hs
@@ -1,0 +1,9 @@
+-- | This module exports all 'Foldable'-related and 'Traversable'-related stuff.
+
+module Universum.Foldable
+       ( module Universum.Foldable.Fold
+       , module Universum.Foldable.Reexport
+       ) where
+
+import Universum.Foldable.Fold
+import Universum.Foldable.Reexport