+++ src/Universum/Monad.hs
@@ -3,14 +3,12 @@
 -- | Reexporting useful monadic stuff.
 
 module Universum.Monad
-       ( module Universum.Monad.Container
-       , module Universum.Monad.Either
+       ( module Universum.Monad.Either
        , module Universum.Monad.Maybe
        , module Universum.Monad.Reexport
        , module Universum.Monad.Trans
        ) where
 
-import Universum.Monad.Container
 import Universum.Monad.Either
 import Universum.Monad.Maybe
 import Universum.Monad.Reexport